diff: Pass 'changes found' flag to external diff tools
When GIT_EXTERNAL_DIFF is defined and git diff is called with any of the
'ignore whitespace' type command line arguments such as -w, -b or
--ignore-space-at-eol, git diff ignores these flags and invokes
GIT_EXTERNAL_DIFF for each changed file in the diff queue. If the diff
queue contains file(s) with only whitespace changes, the external diff
tool displays diffs for changes that the user specifically did not ask
for.

For external diff tools, therefore, it could be useful if git diff set a
GIT_DIFF_FOUND_CHANGES environment variable for each file in the diff
queue. This variable could be examined by the external diff tool to
decide whether to show a diff or skip the file.

To implement it:
(a) Modify diff.c to indicate to external diff tools which file
    contains changes the tool should care about:
        Only when GIT_EXTERNAL_DIFF is defined and either
        --ignore-space-change, --ignore-all-space or
        --ignore-space-at-eol option is set, run diffstat on the current
        file in the diff queue to find out the number of deleted and
        inserted changes. Based on this set the environment variable
        GIT_DIFF_FOUND_CHANGES to "1" or "0". If GIT_EXTERNAL_DIFF is
        defined but none of the 'ignore whitespace' command line
        arguments is set, set GIT_DIFF_FOUND_CHANGES to "1".

(b) Modify git-difftool--helper.sh to make use of the GIT_DIFF_FOUND_CHANGES
    variable: do not invoke the external diff tool on files for which
    GIT_DIFF_FOUND_CHANGES is set to false.

(c) Update documentation and create unit tests.

Signed-off-by: Zoltan Klinger <zoltan.klinger@gmail.com>